Bug fix on connect2network function

+++ dbmodel/utils/fct/gw_fct_connect_to_network.sql
@@ -41,7 +41,9 @@ DECLARE
 	aux_geom public.geometry;
 	v_node_proximity double precision;
 	v_projecttype text;
-
+	v_endfeature_geom public.geometry;
+	v_pjointtype text;
+	v_pjointid text;
 BEGIN
 
     -- Search path
@@ -93,41 +95,64 @@ BEGIN
 		-- compute link
 		IF v_arc.the_geom IS NOT NULL THEN
 
-			IF v_link.userdefined_geom IS TRUE THEN	        
+			IF v_link.userdefined_geom IS TRUE THEN	
+
+				-- get endfeature geometry
+				IF v_link.exit_type='NODE' THEN
+					SELECT node_id, the_geom INTO v_pjointid, v_endfeature_geom FROM node WHERE node_id=v_link.exit_id;
+					v_pjointtype='NODE';
+					
+				ELSIF v_link.exit_type='CONNEC' THEN
+					SELECT pjoint_type, pjoint_id, the_geom INTO v_pjointtype, v_pjointid, v_endfeature_geom FROM connec WHERE connec_id=v_link.exit_id;
+
+				ELSIF v_link.exit_type='GULLY' THEN
+					SELECT pjoint_type, pjoint_id, the_geom INTO v_pjointtype, v_pjointid, v_endfeature_geom FROM gully WHERE gully_id=v_link.exit_id;
+
+				ELSIF v_link.exit_type='VNODE' THEN
+					-- in this case whe don't use the v_link record variable because perhaps there is not link
+					v_endfeature_geom = v_arc.the_geom;
+					v_pjointtype = v_link.exit_type;
+					v_pjointid = v_link.exit_id;
+					
+				END IF;
 
 				-- Reverse (if it's need) the existing link geometry
 				IF (st_dwithin (st_startpoint(v_link.the_geom), v_connect.the_geom, 0.01)) IS FALSE THEN
-					point_aux := St_closestpoint(v_arc.the_geom, St_startpoint(v_link.the_geom));
+					point_aux := St_closestpoint(v_endfeature_geom, St_startpoint(v_link.the_geom));
 					v_link.the_geom = ST_SetPoint(v_link.the_geom, 0, point_aux) ; 
 				ELSE
-					point_aux := St_closestpoint(v_arc.the_geom, St_endpoint(v_link.the_geom));
+					point_aux := St_closestpoint(v_endfeature_geom, St_endpoint(v_link.the_geom));
 					v_link.the_geom = ST_SetPoint(v_link.the_geom, (ST_NumPoints(v_link.the_geom) - 1),point_aux); 
 				END IF;
-			ELSE	
-				v_link.the_geom := ST_ShortestLine(v_connect.the_geom, v_arc.the_geom);
+				
+			ELSE -- in this case only arc is posible (vnode)
+				v_link.the_geom := ST_ShortestLine(v_connect.the_geom, v_endfeature_geom);
 				v_link.userdefined_geom = FALSE;	
 			END IF;
 
 			v_exit.the_geom = ST_EndPoint(v_link.the_geom);
 		END IF;
-
-		-- vnode
-		DELETE FROM vnode WHERE vnode_id=v_exit.vnode_id::integer;
+		
 		IF v_exit.the_geom IS NOT NULL THEN
-			INSERT INTO vnode (vnode_id, vnode_type, state, sector_id, dma_id, expl_id, the_geom) 
-			VALUES ((SELECT nextval('vnode_vnode_id_seq')), 'AUTO', v_arc.state, v_arc.sector_id, v_arc.dma_id, v_arc.expl_id, v_exit.the_geom) RETURNING vnode_id INTO v_exit_id;
-	
-			v_link.exit_id = v_exit_id;
-			v_link.exit_type = 'VNODE';
 
-			-- link
+			-- vnode, only for those links connected to vnode
+			IF v_link.exit_type = 'VNODE' THEN
+				DELETE FROM vnode WHERE vnode_id=v_exit.vnode_id::integer;			
+				INSERT INTO vnode (vnode_id, vnode_type, state, sector_id, dma_id, expl_id, the_geom) 
+				VALUES ((SELECT nextval('vnode_vnode_id_seq')), 'AUTO', v_arc.state, v_arc.sector_id, v_arc.dma_id, v_arc.expl_id, v_exit.the_geom) RETURNING vnode_id INTO v_exit_id;	
+				v_link.exit_id = v_exit_id;
+				v_link.exit_type = 'VNODE';
+			END IF;
+
+			-- link for all links
 			DELETE FROM link WHERE link_id=v_link.link_id;
 			INSERT INTO link (link_id, the_geom, feature_id, feature_type, exit_type, exit_id, userdefined_geom, state, expl_id) 
-			VALUES ((SELECT nextval('link_link_id_seq')), v_link.the_geom, connect_id_aux, feature_type_aux, v_link.exit_type, v_link.exit_id, v_link.userdefined_geom, v_connect.state, v_connect.expl_id);
+			VALUES ((SELECT nextval('link_link_id_seq')), v_link.the_geom, connect_id_aux, feature_type_aux, v_link.exit_type, v_link.exit_id, 
+			v_link.userdefined_geom, v_connect.state, v_connect.expl_id);
 
-			-- Update conntect attributes
+			-- Update connect attributes
 			IF feature_type_aux ='CONNEC' THEN          
-				UPDATE connec SET arc_id=v_connect.arc_id, dma_id=v_connect.dma_id, sector_id=v_connect.sector_id, pjoint_type=v_link.exit_type, pjoint_id=v_link.exit_id
+				UPDATE connec SET arc_id=v_connect.arc_id, dma_id=v_connect.dma_id, sector_id=v_connect.sector_id, pjoint_type=v_pjointtype, pjoint_id=v_pjointid
 				WHERE connec_id = connect_id_aux;
 
 				-- update specific fields for ws projects
@@ -142,7 +167,7 @@ BEGIN
 				END IF;
 			
 			ELSIF feature_type_aux ='GULLY' THEN 
-				UPDATE gully SET arc_id=v_connect.arc_id, dma_id=v_connect.dma_id, sector_id=v_connect.sector_id, pjoint_type=v_link.exit_type, pjoint_id=v_link.exit_id
+				UPDATE gully SET arc_id=v_connect.arc_id, dma_id=v_connect.dma_id, sector_id=v_connect.sector_id, pjoint_type=v_pjointtype, pjoint_id=v_pjointid
 				WHERE gully_id = connect_id_aux;
 
 				-- Update state_type if edit_connect_update_statetype is TRUE
